--- internal/app/apptainer/oci_delete_linux.go
@@ -1,63 +1,0 @@
-// Copyright (c) Contributors to the Apptainer project, established as
-//   Apptainer a Series of LF Projects LLC.
-//   For website terms of use, trademark policy, privacy policy and other
-//   project policies see https://lfprojects.org/policies
-// Copyright (c) 2018-2022, Sylabs Inc. All rights reserved.
-// This software is licensed under a 3-clause BSD license. Please consult the
-// LICENSE.md file distributed with the sources of this project regarding your
-// rights to use or distribute this software.
-
-package apptainer
-
-import (
-	"context"
-	"fmt"
-	"os"
-	"os/exec"
-	"path/filepath"
-
-	"github.com/apptainer/apptainer/internal/pkg/util/bin"
-	"github.com/apptainer/apptainer/pkg/sylog"
-)
-
-// OciDelete deletes container resources
-func OciDelete(ctx context.Context, containerID string) error {
-	runc, err := bin.FindBin("runc")
-	if err != nil {
-		return err
-	}
-	runcArgs := []string{
-		"--root", RuncStateDir,
-		"delete",
-		containerID,
-	}
-
-	cmd := exec.Command(runc, runcArgs...)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	cmd.Stdin = os.Stdout
-	sylog.Debugf("Calling runc with args %v", runcArgs)
-	err = cmd.Run()
-	if err != nil {
-		return fmt.Errorf("while calling runc delete: %w", err)
-	}
-
-	sd, err := stateDir(containerID)
-	if err != nil {
-		return fmt.Errorf("while computing state directory: %w", err)
-	}
-
-	bLink := filepath.Join(sd, bundleLink)
-	bundle, err := filepath.EvalSymlinks(bLink)
-	if err != nil {
-		return fmt.Errorf("while finding bundle directory: %w", err)
-	}
-
-	sylog.Debugf("Removing bundle symlink")
-	if err := os.Remove(bLink); err != nil {
-		return fmt.Errorf("while removing bundle symlink: %w", err)
-	}
-
-	sylog.Debugf("Releasing bundle lock")
-	return releaseBundle(bundle)
-}
