--- internal/app/apptainer/oci_update_linux.go
@@ -1,39 +1,0 @@
-// Copyright (c) Contributors to the Apptainer project, established as
-//   Apptainer a Series of LF Projects LLC.
-//   For website terms of use, trademark policy, privacy policy and other
-//   project policies see https://lfprojects.org/policies
-// Copyright (c) 2018-2022, Sylabs Inc. All rights reserved.
-// This software is licensed under a 3-clause BSD license. Please consult the
-// LICENSE.md file distributed with the sources of this project regarding your
-// rights to use or distribute this software.
-
-package apptainer
-
-import (
-	"os"
-	"os/exec"
-
-	"github.com/apptainer/apptainer/internal/pkg/util/bin"
-	"github.com/apptainer/apptainer/pkg/sylog"
-)
-
-// OciUpdate updates container cgroups resources
-func OciUpdate(containerID string, args *OciArgs) error {
-	runc, err := bin.FindBin("runc")
-	if err != nil {
-		return err
-	}
-	runcArgs := []string{
-		"--root", RuncStateDir,
-		"update",
-		"-r", args.FromFile,
-		containerID,
-	}
-
-	cmd := exec.Command(runc, runcArgs...)
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-	cmd.Stdin = os.Stdout
-	sylog.Debugf("Calling runc with args %v", runcArgs)
-	return cmd.Run()
-}
